+++ app/src/main/java/za/grounded/jwc_app/viewmodels/CartViewModel.java
@@ -23,7 +23,7 @@ public class CartViewModel extends AndroidViewModel {
         return this.cartItemRepository.getCartItemList();
     }
 
-    public void updateCartItemQuantity(int id, int val) {
+    public void updateCartItemQuantity(Long id, Long val) {
         this.cartItemRepository.updateCartItemQuantity(id, val);
     }
 
@@ -31,7 +31,7 @@ public class CartViewModel extends AndroidViewModel {
         return this.cartItemRepository.calculateTotalCartPrice();
     }
 
-    public void deleteCartItem(int id) {
+    public void deleteCartItem(Long id) {
         this.cartItemRepository.deleteCartItem(id);
     }
 }
